Start setuo(Gaby)
Alles toegevoagd zoals de migrations en pagina en models.

Seeder en een foutmelding moet nog gefixt worden

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -31,7 +31,17 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('users');
-        $table->dropColumn('role');
+        if (Schema::hasTable('users')) {
+            Schema::table('users', function (Blueprint $table) {
+                if (Schema::hasColumn('users', 'role')) {
+                    $table->dropColumn('role');
+                }
+            });
+    
+            Schema::dropIfExists('users');
+        }
     }
+    
+    
 };
+
